pinecone-demo/requirements.txt
pinecone-client
numpy
pandas
requests